Add new screen to navigation
